nvim-rs: Switch back to my fork
Looks like the neovim API has gotten a bit out of date in nvim-rs, and as a
result we are seeing random crashes due to the return arguments of
functions not matching what they should. Fix that by updating it ourselves
in our own fork and switch neovim-gtk over to using that.